+++ routers/index.js
@@ -35,19 +35,19 @@ module.exports = _ => {
   /*
    *  用户模块
    */
-  _.get("/user/list", user.list); // 用户信息列表
+  _.get("/user/list", authentication, user.list); // 用户信息列表
   _.get("/user/:mail/loginPermit", user.permit); // 登录申请
   _.post("/user/login", user.login); // 登录
 
   /*
    *  工作流模块
    */
-  _.get("/workflow/list", workflow.list); // 工作流信息列表
+  _.get("/workflow/list", authentication, workflow.list); // 工作流信息列表
 
   /*
    *  文件系统模块
    */
-  _.get("/fs/list", goodses.list); // 更新文件信息列表
+  _.get("/fs/list", authentication, goodses.list); // 更新文件信息列表
 
   /*
    *  测试：开发者信息